+++ legacy/keystore/keystore.go
@@ -41,7 +41,7 @@ import (
 	"github.com/btcsuite/btcd/txscript"
 	"github.com/btcsuite/btcd/wire"
 	"github.com/btcsuite/btcutil"
-	"github.com/btcsuite/btcwallet/rename"
+	"github.com/btcsuite/btcwallet/legacy/rename"
 )
 
 const (
